+++ apps/angular-lib-dev/src/app/dropdown/dropdown.component.ts
@@ -12,74 +12,87 @@ import { DropdownOption, DropdownTexts } from '@sebgroup/extract'
 export class DropdownComponent {
   texts: DropdownTexts = {
     placeholder: 'Select',
+    searchPlaceholder: 'Search',
   }
 
-  options: Array<DropdownOption> = [
+  options: DropdownOption[] = [
     {
-      key: 'Tacos',
+      label: 'Tacos',
       value: 'tacos',
     },
     {
-      key: 'Pizza',
+      label: 'Pizza',
       value: 'pizza',
     },
     {
-      key: 'Sushi',
+      label: 'Sushi',
       value: 'sushi',
     },
   ]
 
-  optionsWithInitial = [
+  optionsWithInitial: DropdownOption[] = [
     {
-      key: 'Tacos',
+      label: 'Tacos',
       value: 'tacos',
     },
     {
-      key: 'Pizza',
+      label: 'Pizza',
       value: 'pizza',
       selected: true,
     },
     {
-      key: 'Sushi',
+      label: 'Sushi',
       value: 'sushi',
     },
   ]
 
-  customOptionsInitial = [
+  customOptionsInitial: DropdownOption[] = [
     {
       name: 'Tacos',
-      val: 'tacos',
-      kitchen: 'mexican',
+      val: {
+        id: 'tacos',
+        kitchen: 'mexican',
+      },
     },
     {
       name: 'Pizza',
-      val: 'pizza',
+      val: {
+        id: 'pizza',
+        kitchen: 'italian',
+      },
       selected: true,
-      kitchen: 'italian',
     },
     {
       name: 'Sushi',
-      val: 'sushi',
+      val: {
+        id: 'sushi',
+        kitchen: 'japanese',
+      },
       selected: true,
-      kitchen: 'japanese',
     },
   ]
 
-  customOptions = [
+  customOptions: DropdownOption[] = [
     {
       name: 'Tacos',
-      val: 'tacos',
-      kitchen: 'mexican',
+      val: {
+        id: 'tacos',
+        kitchen: 'mexican',
+      },
     },
     {
       name: 'Pizza',
-      val: 'pizza',
-      kitchen: 'italian',
+      val: {
+        id: 'pizza',
+        kitchen: 'italian',
+      },
     },
     {
       name: 'Sushi',
-      val: 'sushi',
-      kitchen: 'japanese',
+      val: {
+        id: 'sushi',
+        kitchen: 'japanese',
+      },
     },
   ]
 
@@ -178,18 +191,18 @@ export class DropdownComponent {
     otherCountries: [null, [Validators.required, Validators.minLength(2)]],
   })
 
-  dropdownValue0: any = undefined
-  dropdownValue1: any = undefined
-  dropdownValue2 = 'sushi'
+  dropdownValue0: string | undefined = 'sushi'
+  dropdownValue1: string | undefined
+  dropdownValue2: string | undefined = 'sushi'
 
-  dropdownMultiValue0: any = undefined
-  dropdownMultiValue1: any = undefined
-  dropdownMultiValue2 = 'sushi'
+  dropdownMultiValue0: string[] | undefined = undefined
+  dropdownMultiValue1: unknown[] | undefined = undefined
+  dropdownMultiValue2: unknown[] | undefined = [this.customOptions[2].val]
 
-  dropdownValue3: any = undefined
-  dropdownValue4 = 'tacos'
-  dropdownValue5 = 'pizza'
-  dropdownValue6 = undefined
+  dropdownValue3: string | undefined = undefined
+  dropdownValue4: string | undefined = 'tacos'
+  dropdownValue5: unknown[] | undefined = this.customOptions[1].val
+  dropdownValue6: unknown[] | undefined = undefined
 
   constructor(private _fb: FormBuilder) {}
 
@@ -198,7 +211,15 @@ export class DropdownComponent {
     console.log(this.validationForm.value)
   }
 
-  test($event: any) {
-    console.log($event)
+  test($event: Event) {
+    console.log('Test', $event)
+  }
+
+  compareWith(o1: { id: string }, o2: { id: string }) {
+    return o1.id === o2.id
+  }
+
+  searchFilter(search: string, value: { id: string; kitchen: string }) {
+    return value.kitchen.toLowerCase().includes(search.toLowerCase())
   }
 }